+++ src/main/java/com/javaweb/repository/UserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<UserEntity, Long> , UserRepositoryCustom {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> , UserRepositoryCustom {
     UserEntity findOneByUsernameAndStatus(String name, int status);
     Page<UserEntity> findByUsernameContainingIgnoreCaseOrFullNameContainingIgnoreCaseAndStatusNot(String username, String fullName, int status,
                                                                                                   Pageable pageable);
